refactor: delete MUIHeader
