+++ src/setupTests.ts
@@ -1,8 +1,6 @@
 import '@testing-library/jest-dom';
 import { createRoot } from 'react-dom/client';
 
-// import { unmountComponentAtNode } from 'react-dom';
-
 let container: Element | null;
 beforeEach(() => {
   container = document.createElement('div');
@@ -11,7 +9,6 @@ beforeEach(() => {
 
 afterEach(() => {
   if (container) {
-    // unmountComponentAtNode(container);
     const root = createRoot(container);
     root.unmount();
     container.remove();
